refactor: move routing primitives to shared
