--- src/features/navigate/index.ts
@@ -1,2 +1,0 @@
-export { routes } from "./model/routes";
-export { useNavigation } from "./model/useNavigation";
