+++ src/shared/router/index.ts
@@ -1,0 +1,2 @@
+export { routes } from "./routes";
+export { useNavigation } from "./useNavigation";
